requirements.txt
fitz==0.0.1.dev2
langchain==0.3.24
langchain_core==0.3.56
langchain_google_genai==2.1.3
langgraph==0.3.34
pinecone==6.0.2
python-dotenv==1.1.0
sentence_transformers==4.1.0
streamlit==1.44.1
typing_extensions==4.13.2
